+++ CHANGELOG.md
@@ -15,6 +15,7 @@ use [Keep a Changelog] (http://keepachangelog.com/).
 
 - `ArrayHelper::column()` with non-Closure `$index_key` was incorrectly assigning to `$finalValue` instead of `$finalIndex`
 - `FileHelper::uniformizePathSeparator()` was removing double slashes instead of collapsing them, corrupting paths
+- `b_array_simple()` bootstrap function was ignoring the `$prefix` parameter due to accidental reassignment to `null`
 
 ## [1.12.0] - 2025-05-27